src/test/java/com/judalabs/leetcode/topswe/P026_200Test.java
package com.judalabs.leetcode.topswe;

import java.util.stream.Stream;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

class P026_200Test {

    @MethodSource("examples")
    @ParameterizedTest
    void tests(char[][] input, int output) {
        Assertions.assertEquals(output, new P026_200().numIslands(input));
    }

    public static Stream<Arguments> examples() {
        return Stream.of(
                Arguments.of(new char[][]{
                        {'1', '1', '1', '1', '0'},
                        {'1', '1', '0', '1', '0'},
                        {'1', '1', '0', '0', '0'},
                        {'0', '0', '0', '0', '0'}
                }, 1),
                Arguments.of(new char[][]{
                        {'1', '1', '0', '0', '0'},
                        {'1', '1', '0', '0', '0'},
                        {'0', '0', '1', '0', '0'},
                        {'0', '0', '0', '1', '1'}
                }, 3)
        );
    }

}